weather_pipeline/requirements.txt
requests
pandas
pyyaml
pydantic
pyarrow